requirements.txt
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
Django==5.0
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
Pillow==10.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python3-xlib==0.15
pytweening==1.0.7
PyVirtualDisplay==3.0
sqlparse==0.4.4
Werkzeug==3.0.1
